pyright fixes

+++ meshio/vtk/_vtk.py
@@ -341,7 +341,9 @@ def _check_mesh(info):
 
 def _generate_cells(dim):
     ele_dim = [d - 1 for d in dim if d > 1]
-    ele_no = numpy.prod(ele_dim, dtype=int)
+    # TODO use math.prod when requiring Python 3.8+? this would save the int conversion
+    # <https://github.com/microsoft/pyright/issues/1226>
+    ele_no = int(numpy.prod(ele_dim))
     spatial_dim = len(ele_dim)
 
     if spatial_dim == 1:
